Azure OpenAI improvements and fixes; Improve Graphiti Azure OpenAI config
